Split autocomplete Playwright tests by variant and expand e2e coverage
